% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{generate_data}
\alias{generate_data}
\title{: Generating panel data for heterogeneity testing}
\usage{
generate_data(N, tt, beta, sigma)
}
\arguments{
\item{N:}{Number of individuals}

\item{tt:}{Number of time periods}

\item{beta:}{Coefficients}

\item{sigma:}{Standard deviation}
}
\value{
: A list of data
}
\description{
: Generate panel data for heterogeneity testing
}
