% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{hetero_probit}
\alias{hetero_probit}
\title{: Rcpp version for the heterogenous probit model}
\usage{
hetero_probit(df)
}
\arguments{
\item{df:}{Data frame}
}
\value{
: A list of coefficients
}
\description{
: Rcpp version for the heterogenous probit model
}
