\name{HomogeneityTest-package}
\alias{HomogeneityTest-package}
\alias{HomogeneityTest}
\docType{package}
\title{
\packageTitle{HomogeneityTest}
}
\description{
\packageDescription{HomogeneityTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HomogeneityTest}
\packageIndices{HomogeneityTest}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{HomogeneityTest}

Maintainer: \packageMaintainer{HomogeneityTest}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\examples{
~~ simple examples of the most important functions ~~
}
