% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{probit_glm}
\alias{probit_glm}
\title{: Rcpp version for glm probit}
\usage{
probit_glm(X, Y, max_iter = 1000, tol = 1e-06)
}
\arguments{
\item{X:}{Design matrix}

\item{Y:}{Response variable}

\item{max_iter:}{Maximum number of iterations}

\item{tol:}{Tolerance}
}
\value{
: A list of coefficients
}
\description{
: Rcpp version for glm probit
}
