% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{quantile_function}
\alias{quantile_function}
\title{: Rcpp version for the heterogenous probit model}
\usage{
quantile_function(x, q)
}
\arguments{
\item{df:}{Data frame}
}
\value{
: A list of coefficients
}
\description{
: Rcpp version for the heterogenous probit model
}
